# master-rsi-backend/app/Models/Image.php
<?php

// namespace App\Models;

// use Illuminate\Database\Eloquent\Model;

// class Image extends Model
// {
//     //
// }
namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Image extends Model
{
    use HasFactory;

    /**
     * The table associated with the model.
     *
     * @var string
     */
    protected $table = 'images';

    /**
     * The primary key for the model.
     *
     * @var string
     */
    protected $primaryKey = 'id';
    public $timestamps = false;

    /**
     * The attributes that are mass assignable.
     *
     * @var array
     */
    protected $fillable = [
        'name',
        'type',
        'size',
        'bin_img'
    ];

    /**
     * The attributes that should be cast.
     *
     * @var array
     */
    protected $casts = [
        'size' => 'integer'
    ];

    /**
     * Get the image URL attribute.
     *
     * @return string
     */
    public function getUrlAttribute()
    {
        return 'data:image/' . $this->type . ';base64,' . $this->bin_img;
    }

    /**
     * Scope a query to only include images of a certain type.
     *
     * @param  \Illuminate\Database\Eloquent\Builder  $query
     * @param  string  $type
     * @return \Illuminate\Database\Eloquent\Builder
     */
    public function scopeOfType($query, $type)
    {
        return $query->where('type', $type);
    }
}